requirements.txt
mediapipe==0.10.14
opencv-python
pyserial
